+++ packages/gatsby-plugin-manifest/src/gatsby-node.js
@@ -17,34 +17,88 @@ try {
   // doesn't support cpu-core-count utility.
 }
 
-function generateIcons(icons, srcIcon) {
-  return Promise.all(
-    icons.map(async icon => {
-      const size = parseInt(
-        icon.sizes.substring(0, icon.sizes.lastIndexOf(`x`))
-      )
-      const imgPath = path.join(`public`, icon.src)
+async function generateIcon(icon, srcIcon) {
+  const imgPath = path.join(`public`, icon.src)
 
-      // For vector graphics, instruct sharp to use a pixel density
-      // suitable for the resolution we're rasterizing to.
-      // For pixel graphics sources this has no effect.
-      // Sharp accept density from 1 to 2400
-      const density = Math.min(2400, Math.max(1, size))
+  // console.log(`generating icon: `, icon.src)
+  // if (fs.existsSync(imgPath)) {
+  //   console.log(`icon already Exists, not regenerating`)
+  //   return true
+  // }
+  const size = parseInt(icon.sizes.substring(0, icon.sizes.lastIndexOf(`x`)))
 
-      return sharp(srcIcon, { density })
-        .resize({
-          width: size,
-          height: size,
-          fit: `contain`,
-          background: { r: 255, g: 255, b: 255, alpha: 0 },
-        })
-        .toFile(imgPath)
+  // For vector graphics, instruct sharp to use a pixel density
+  // suitable for the resolution we're rasterizing to.
+  // For pixel graphics sources this has no effect.
+  // Sharp accept density from 1 to 2400
+  const density = Math.min(2400, Math.max(1, size))
+
+  return sharp(srcIcon, { density })
+    .resize({
+      width: size,
+      height: size,
+      fit: `contain`,
+      background: { r: 255, g: 255, b: 255, alpha: 0 },
     })
-  )
+    .toFile(imgPath)
 }
 
-exports.onPostBootstrap = async ({ reporter }, pluginOptions) => {
+async function checkCache(cache, icon, srcIcon, srcIconDigest, callback) {
+  const cacheKey = createContentDigest(`${icon.src}${srcIcon}${srcIconDigest}`)
+
+  let created = cache.get(cacheKey, srcIcon)
+
+  if (!created) {
+    cache.set(cacheKey, true)
+
+    try {
+      // console.log(`creating icon`, icon.src, srcIcon)
+      await callback(icon, srcIcon)
+    } catch (e) {
+      cache.set(cacheKey, false)
+      throw e
+    }
+  } else {
+    // console.log(`icon exists`, icon.src, srcIcon)
+  }
+}
+
+exports.onPostBootstrap = async ({ reporter }, { localize, ...manifest }) => {
+  const activity = reporter.activityTimer(`Build manifest and related icons`)
+  activity.start()
+
+  let cache = new Map()
+
+  await makeManifest(cache, reporter, manifest)
+
+  if (Array.isArray(localize)) {
+    const locales = [...localize]
+    await Promise.all(
+      locales.map(locale => {
+        let cacheModeOverride = {}
+
+        /* localization requires unique filenames for output files if a different src Icon is defined.
+           otherwise one language would override anothers icons in automatic mode.
+        */
+        if (locale.hasOwnProperty(`icon`) && !locale.hasOwnProperty(`icons`)) {
+          // console.debug(`OVERRIDING CACHE BUSTING`, locale)
+          cacheModeOverride = { cache_busting_mode: `name` }
+        }
+
+        return makeManifest(cache, reporter, {
+          ...manifest,
+          ...locale,
+          ...cacheModeOverride,
+        })
+      })
+    )
+  }
+  activity.end()
+}
+
+const makeManifest = async (cache, reporter, pluginOptions) => {
   const { icon, ...manifest } = pluginOptions
+  const suffix = pluginOptions.lang ? `_${pluginOptions.lang}` : ``
 
   // Delete options we won't pass to the manifest.webmanifest.
   delete manifest.plugins
@@ -55,12 +109,9 @@ exports.onPostBootstrap = async ({ reporter }, pluginOptions) => {
   delete manifest.icon_options
   delete manifest.include_favicon
 
-  const activity = reporter.activityTimer(`Build manifest and related icons`)
-  activity.start()
-
   // If icons are not manually defined, use the default icon set.
   if (!manifest.icons) {
-    manifest.icons = defaultIcons
+    manifest.icons = [...defaultIcons]
   }
 
   // Specify extra options for each icon (if requested).
@@ -111,30 +162,38 @@ exports.onPostBootstrap = async ({ reporter }, pluginOptions) => {
         ? pluginOptions.cache_busting_mode
         : `query`
 
+    const iconDigest = createContentDigest(fs.readFileSync(icon))
+
     //if cacheBusting is being done via url query icons must be generated before cache busting runs
     if (cacheMode === `query`) {
-      await generateIcons(manifest.icons, icon)
+      await Promise.all(
+        manifest.icons.map(dstIcon =>
+          checkCache(cache, dstIcon, icon, iconDigest, generateIcon)
+        )
+      )
     }
 
     if (cacheMode !== `none`) {
-      const iconDigest = createContentDigest(fs.readFileSync(icon))
-
-      manifest.icons.forEach(icon => {
-        icon.src = addDigestToPath(icon.src, iconDigest, cacheMode)
+      manifest.icons = manifest.icons.map(icon => {
+        let newIcon = { ...icon }
+        newIcon.src = addDigestToPath(icon.src, iconDigest, cacheMode)
+        return newIcon
       })
     }
 
     //if file names are being modified by cacheBusting icons must be generated after cache busting runs
     if (cacheMode !== `query`) {
-      await generateIcons(manifest.icons, icon)
+      await Promise.all(
+        manifest.icons.map(dstIcon =>
+          checkCache(cache, dstIcon, icon, iconDigest, generateIcon)
+        )
+      )
     }
   }
 
   //Write manifest
   fs.writeFileSync(
-    path.join(`public`, `manifest.webmanifest`),
+    path.join(`public`, `manifest${suffix}.webmanifest`),
     JSON.stringify(manifest)
   )
-
-  activity.end()
 }